Improves Player class

+++ src/entities/Entity.java
@@ -4,7 +4,7 @@ public abstract class Entity {
 
     protected float x, y;
 
-    public Entity(int x, int y){
+    public Entity(float x, float y){
         this.x = x;
         this.y = y;
     }
